Separate key_buffer struct.
